AGS: Fix some warnings (thanks [md5]).

+++ engines/ags/script.cpp
@@ -619,7 +619,7 @@ void ccInstance::runCodeFrom(uint32 start) {
 					      "bytes of %d) on line %d",
 					      int1, int2, _lineNumber);
 				// FIXME: check bounds
-				for (uint i = 0; i < int1; ++i)
+				for (uint i = 0; i < (uint) int1; ++i)
 					_stack[i + _registers[SREG_MAR]._value].invalidate();
 				break;
 			}
